Bump `storybook` to v.6.5.15 and regenerate React project. Bump React to latest version (v.8).
Removed `storybook` from project and then reinstalled it via `npx storybook upgrade`.
This was done to fix current installation problems that prevented the storybook build
from succeeding. This induced an additional breaking issue with typings, though,
leading to several components in the project throwing a `Not a JSX.Element` error.
Followed `https://bobbyhadz.com/blog/react-cannot-be-used-as-a-jsx-component#update-your-reactjs-typings`
to update typings to latest version, solving the problem with `Component` not
being recognized as a proper `JSX.Element`. This leads to `React.FC<PropsType>` not
adding `children` as a default parameter, so as a last step, replaced faulty
occurrences of `React.FC` to plan functions including `children?: React.ReactNode;`
to their props declarations.